Maybe avoid TSC freaking out.
