+++ js/hang/src/worklet/render.ts
@@ -1,23 +1,4 @@
-export type Message = Init | Data;
-
-export interface Data {
-	type: "data";
-	data: Float32Array[];
-	timestamp: number;
-}
-
-export interface Init {
-	type: "init";
-	sampleRate: number;
-	channelCount: number;
-	latency: DOMHighResTimeStamp;
-}
-
-export interface Status {
-	type: "status";
-	available: number;
-	utilization: number;
-}
+import type { Data, Init, Message, Status } from ".";
 
 class Render extends AudioWorkletProcessor {
 	#buffer: Float32Array[] = [];
