+++ js/hang/tsconfig.json
@@ -1,7 +1,7 @@
 {
 	"extends": "../tsconfig.json",
 	"include": ["src"],
-	"exclude": ["node_modules", "dist"],
+	"exclude": ["node_modules", "dist", "src/worklet"],
 	"compilerOptions": {
 		"outDir": "dist",
 		"jsx": "preserve",
